rearrange smatrix modules
